requirements.txt
clip==1.0
faiss_cpu==1.8.0.post1
faiss_gpu==1.7.2
matplotlib==3.7.5
numpy==1.24.3
numpy==1.24.4
opencv_python==4.10.0.84
Pillow==9.5.0
Pillow==11.0.0
scikit_learn==1.3.2
scipy==1.14.1
segment_anything==1.0
timm==0.3.2
torch==2.2.0
torchvision==0.17.0
torchvision==0.18.1
tqdm==4.65.2
